feature(openapi): Ensure validating error responses
Previously validation of request / response done via `openapi_operation`
decorator. But this provides several limitations, most notably of,

- It is very hard to validate error responses as errors from decorator
  or handler catched by `error_middleware` and response from that
  middleware do not validated over OpenAPI schema
- There is small possibility in supporting class based views properly
  via `openapi_operation` decorator, as class based views is not
  regular functions and should be treat differently

With that in mind it was decided to entirely remove `@openapi_operation`
and provide `openapi_middleware` instead. With addition of
`get_error_response` coroutine to `aiohttp-middlewares==1.1.0` now it
is possible to validate even error responses and it should be possible
to support class based views as well.

Issue: #45
Fixes: #26